Move tests for multipart form data directly into their own class
